dynamically handling the column names in excel load
dynamically handled the column names in the excel file irrespective of
the number and order of the columns. uses the first row as column
headers.
1) added the Load() function to the helpers\handlers.go to for this
purpose. which load the given excel file (that is on server) into the
data structure of type []map[string]string.
2) and created new route "/upload" and new route handler function
handleUpld() that calls the Load() function and use the returned dataset
to Marshal and send as response.